Modify struct vattr: 	Increase va_nlink, va_fileid (the inode number), and va_gen from 	32 bit to 64 bit integers.
	Add va_uid_uuid, va_gid_uuid, and va_fsid_uuid, and flags to
	indicate that these fields are valid.  The original va_uid and
	va_gid are retained.

	This change has no external visibility.

Modify struct statvfs:
	Use spare fields to add f_fsid_uuid and f_uid_uuid to the
	structure, and flags indicating that those fields are valid.

	This change has minimal external visibility. The size of the
	structure has not changed.

Modify struct stat:
	Add a new file type S_IFDB.  DB files are like regular files but
	access data on a record by record basis.  The seek position is a
	64 bit record key and not a byte offset.  Further work in this
	area will be done later on to support related UIO operations.

	This change has minimal external visibility. The size of the
	structure has not changed.

+++ sys/kern/vfs_cache.c
@@ -67,7 +67,7 @@
  *
  *	@(#)vfs_cache.c	8.5 (Berkeley) 3/22/95
  * $FreeBSD: src/sys/kern/vfs_cache.c,v 1.42.2.6 2001/10/05 20:07:03 dillon Exp $
- * $DragonFly: src/sys/kern/vfs_cache.c,v 1.84 2007/08/13 17:43:55 dillon Exp $
+ * $DragonFly: src/sys/kern/vfs_cache.c,v 1.85 2007/11/02 19:52:25 dillon Exp $
  */
 
 #include <sys/param.h>
@@ -1498,7 +1498,7 @@ cache_inefficient_scan(struct nchandle *nch, struct ucred *cred,
 	if ((error = cache_vref(nch, cred, &pvp)) != 0)
 		return (error);
 	if (ncvp_debug)
-		kprintf("inefficient_scan: directory iosize %ld vattr fileid = %ld\n", vat.va_blocksize, (long)vat.va_fileid);
+		kprintf("inefficient_scan: directory iosize %ld vattr fileid = %lld\n", vat.va_blocksize, vat.va_fileid);
 	if ((blksize = vat.va_blocksize) == 0)
 		blksize = DEV_BSIZE;
 	rbuf = kmalloc(blksize, M_TEMP, M_WAITOK);
@@ -1533,7 +1533,7 @@ again:
 			    den->d_ino == vat.va_fileid) {
 				if (ncvp_debug) {
 					kprintf("cache_inefficient_scan: "
-					       "MATCHED inode %ld path %s/%*.*s\n",
+					       "MATCHED inode %lld path %s/%*.*s\n",
 					       vat.va_fileid, nch->ncp->nc_name,
 					       den->d_namlen, den->d_namlen,
 					       den->d_name);
